CodeGen: Parse* can return a value (#2683)
Previously, all Parse* methods returne void. That made it hard to generate a parser that composes larger objects from objects generated by children elements.

This change adds two features that allow to use this "composing" approach.

The first one is addition of mapping between XSD complex type and C# type. This mapping means rest of the code can expect Parse* method to return a value and that value is stored and passed to the hook. The precise type might differ, e.g. when there is a sequence, use List<>.

The other change modifies generated Parse* methods to return a value. The idea is that the hook will receive all values in the element and does the conversion. Because hook has to return value, it can no longer be a partial method for mapped types. It remains a partial method for unmapped complex types.