Add a stub for cell formats
This should enventually grow to fully replace current XLStyle and
XLStylized classes. It will use the XLWorkbookStyles class instead of
current repositories. The internal structure is slighlty more complex,
but should provide support for all use cases:

1. Set a format of various container:
   * one cell (IXLCell)/row (IXLRow)/column (IXLColumn)
   * multiple cells(IXLCells)/rows(IXLRows)/column(IXLColumns)
   * sheet (IXLWorksheet) and
   * area (IXLRangeBase)

3. Should get a format of various containers (see above)

2. Should work even if container is not yet formatted (e.g.
   non-materilized cell or a column), without eager materialization

Formatting is not described anywhere and Excel is kind of insane, but
I think this should work (once properly implemented) to match Excel and
keep the current API (though I am unhappy with the IXLStyle name, it
should be IXFormat or something like that, style is a named format).

The actual code is in a limbo and not yet used anywhere.