Make enum mapper bidictionary
The enum mappers have been used only in readers, to determine enum
value for a text in XML. In order to use it in XML part writers,
the opposite method to get a text for an enum is also necessary.

Writer is not yet done, this is for the future XmlTreeWriter.