Make equality comparer reusable
This will be useful in format setter. The plan is to not run format
modification (e.g., modify format so it has font size 15) for each
cell, but to cache modifications.

The cache will be a dictionary of cellFormat-> cellFormat. The same
input should result in same output, but even classic equals is
unnecessarily slow. Thus dictionary will use reference comparer -
the one separated here.