XLBookPoint should use sheet name
The XLBookPoint used to contain SheetId identifier instead of sheet
name. The original intent was to be able to survive many sheet names
without need to change sheet point.

This turned out not to be such a great choice. The "feature" was never
actually used and internal structures using a book point still has to be
updated when sheet is deleted. Also, it's kind of hard to understand
in relation to other coordinate structures that use sheet name directly.

Therefore the SheetId is switched to SheetName.