Add XLCellFormat format resolution
The XLCellFormat is supposed to be used by a wide variety of range API
objects. As such, it must be able to *set* things across multiple
worksheets.

The *get* part is kind of more complicated. There are two types of API
objects:
* One range API object (IXLColumn, IXLRow, IXLWorksheet, IXLCell...)
* Group range API objects (IXLColumns, IXLCells..)

In the long run, the multi-group range API objects should provide
set-only API instead of get-set API. But for now, the XLCellFormat must
be able to replace the XLStyle. The XLStyle returns only one value.

The added logic to the XLCellFormat enables a resolution of format
through format hierarchy/fallbacks. It requires careful setting of
individual fallbacks. As such, it is a private struct that will be
created through XLCellFormat factory methods.