Add IXLStyle interface to XLCellFormat
Style rework needs to replace current XLStylizedbase+repositories+
XLStyle. The key part is to make the new API object XLCellFormat
implement IXLStyle, so it can be returned in the various Style
properties (IXLCell.Style and so on).

This commit adds the interface IXLStyle and some basic font properties
to the XLCellFormat. The interface was implemented explicitely in
separate files to make a clean separation. The IXLStyle/IXLFont has
many properties with same name but different type as the XLFontFormat.

The XLCells was used as a test range API object, it doesn't represent
internal structure and is thus simpler. The downside is that getters
return sheet properties, but that is a legacy behavior.