Style rework: Fix columns inside borders
The inside borders in Excel (and in general) means set border
to everything but lines that touch outside border.

When IXLStyle was setting inside border to columns, it only set horizontal
(top and bottom) borders. That works when there is only one column, but
not when there are two or more.

This change fixes the problem as part of style-rework. It basically
detects colspans and sets left column, center columns and right column.
There is a special path for one-column, because it doesn't fit into left/
center/right border changes.