Add a representation of ST_Sqref
ST_Sqref is a sequence of references (only within a worksheet, not
global). This type is used in several places, e.g. data validations,
conditional format or protected ranges.

The XLAreaList is a represenation of this type. The XLAreaList also
contains methods for reacting to structural changes. That should be
reusable across several types in a worksheet.

The key difference between single-area objects (ST_CellRef) and
ST_Sqref is that ST_Sqref can be split (e.g. a square can have its
center deleted). That provides a more accurate shifting behavior.

For many OOXML objects that used ST_Sqref, ClosedXML supported only
insert of whole columns or rows, but not of individual ranges that
might cause split. This structure will expand this behavior, so it's
possible to insert/delete a only an area and shift will work properly.

Note: Shift behavior doesn't behave exactly as Excel, but should deal
with all cases without useless splits.